refactor: remove terrainblending special hooks
