+++ include/FrameAnnotations.h
@@ -4,5 +4,4 @@ namespace FrameAnnotations
 {
 	void OnPostPostLoad();
 	void OnDataLoaded();
-	bool IsInRenderShadowmasksPhase();
 }
