+++ src/Hooks.cpp
@@ -12,7 +12,6 @@
 
 #include "Features/InteriorSun.h"
 #include "Features/LightLimitFix.h"
-#include "Features/TerrainBlending.h"
 #include "Features/TerrainHelper.h"
 #include "Features/Upscaling.h"
 #include "Features/VR.h"
@@ -118,15 +117,12 @@ bool Hooks::BSShader_BeginTechnique::thunk(RE::BSShader* shader, uint32_t vertex
 {
 	auto state = globals::state;
 	auto shaderCache = globals::shaderCache;
-	const auto callerRva = static_cast<uint32_t>(reinterpret_cast<std::uintptr_t>(_ReturnAddress()) - REL::Module::get().base());
 
 	state->updateShader = true;
 	state->currentShader = shader;
 
 	state->currentVertexDescriptor = vertexDescriptor;
 	state->currentPixelDescriptor = pixelDescriptor;
-
-	globals::features::terrainBlending.OnBeginTechnique(shader, pixelDescriptor, callerRva);
 
 	state->permutationData.VertexShaderDescriptor = vertexDescriptor;
 	state->permutationData.PixelShaderDescriptor = pixelDescriptor;
@@ -317,9 +313,7 @@ HRESULT WINAPI hk_D3D11CreateDeviceAndSwapChain(
 
 void Hooks::BSGraphics_SetDirtyStates::thunk(bool isCompute)
 {
-	const auto callerRva = static_cast<uint32_t>(reinterpret_cast<std::uintptr_t>(_ReturnAddress()) - REL::Module::get().base());
 	func(isCompute);
-	globals::features::terrainBlending.OnSetDirtyStates(isCompute, callerRva);
 	globals::state->Draw();
 }
